+++ packages/syside-languageserver/src/model/SysML/relationships/expose.ts
@@ -1,5 +1,5 @@
 /********************************************************************************
- * Copyright (c) 2022-2023 Sensmetry UAB and others
+ * Copyright (c) 2022-2025 Sensmetry UAB and others
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -16,6 +16,7 @@
 
 import { AstNode, LangiumDocument } from "langium";
 import { Expose } from "../../../generated/ast";
+import { Visibility } from "../../../utils";
 import { enumerable } from "../../../utils";
 import { Importable, ImportMeta, ImportOptions } from "../../KerML/relationships/import";
 import { ElementIDProvider, metamodelOf, MetatypeProto } from "../../metamodel";
@@ -23,6 +24,21 @@ import { ViewUsageMeta } from "../view-usage";
 
 @metamodelOf(Expose, "abstract")
 export abstract class ExposeMeta<T extends Importable = Importable> extends ImportMeta<T> {
+    /**
+     * Visibility of the element at the end of this import
+     * @see {@link element}
+     */
+    @enumerable
+    override get visibility(): Visibility {
+        return Visibility.protected;
+    }
+
+    override set visibility(value: Visibility) {
+        if (value != Visibility.protected) {
+            throw new Error("Expose visibility must be protected");
+        }
+    }
+
     @enumerable
     override get importsAll(): boolean {
         return true;